requirements.txt
psycopg2-binary
python-dotenv
sqlalchemy
flask
passlib
argon2-cffi